refactor(client): move auth slice into features/auth
