+++ client/src/features/auth/components/ProtectedRoute.tsx
