+++ client/src/pages/Deliveries.tsx
@@ -25,7 +25,7 @@ import DeliveryFormDialog from '../components/delivery/DeliveryFormDialog';
 import DeliveryTimelineDialog from '../components/delivery/DeliveryTimelineDialog';
 import ShippingCompaniesDialog from '../components/delivery/ShippingCompaniesDialog';
 import { formatDateTime, formatCurrency } from '../shared/lib/utils';
-import { useAuthStore } from '../store/authStore';
+import { useAuthStore } from '../features/auth/store/authStore';
 import { useTranslation } from '../shared/i18n/index';
 import { resource } from '../shared/lib/resource';
 import { useApiQuery } from '../shared/lib/apiQuery';
